Added Character Ability DB Services
Merchants now sell Base Ability types and Ability Events.

Added CharacterAbilityService. Abilities are now saved to the database.
Added CharacterKnownAbilityService. Known abilities are now saved to the database.
Added Character Ability synchronization to the Client upon connect.

+++ FishMMO-Unity/Assets/Scripts/Server/Database/Services/Scene/Character/CharacterFriendService.cs
@@ -81,12 +81,15 @@ namespace FishMMO.Server.DatabaseServices
 		/// <summary>
 		/// Removes all characters from a friend list.
 		/// </summary>
-		public static void Delete(NpgsqlDbContext dbContext, long characterID)
+		public static void Delete(NpgsqlDbContext dbContext, long characterID, bool keepData = false)
 		{
-			var characterFriends = dbContext.CharacterFriends.Where(a => a.CharacterID == characterID);
-			if (characterFriends != null)
+			if (!keepData)
 			{
-				dbContext.CharacterFriends.RemoveRange(characterFriends);
+				var characterFriends = dbContext.CharacterFriends.Where(a => a.CharacterID == characterID);
+				if (characterFriends != null)
+				{
+					dbContext.CharacterFriends.RemoveRange(characterFriends);
+				}
 			}
 		}
